move python code to cpp for dim_gather, gather_nd (#6887)
* py2cpp: dim_gather

* fix axis=>dim

* remove py2cpp: gather_nd (#6889)

* Update __init__.py

remove from oneflow.nn.modules.floor import floor_op as floor
