+++ python/oneflow/framework/tensor.py
@@ -729,6 +729,10 @@ def _to(self, *args, **kwargs):
     return flow._C.to(self, *args, **kwargs)
 
 
+def _gather(self, dim, index):
+    return flow._C.dim_gather(self, dim, index, False)
+
+
 def RegisterMethods():
     Tensor.__mul__ = lambda self, other: self.mul(other)
     Tensor.__rmul__ = lambda self, other: self.mul(other)
@@ -871,6 +875,7 @@ def RegisterMethods():
     Tensor.unsqueeze = _unsqueeze
     Tensor.permute = _permute
     Tensor.to = _to
+    Tensor.gather = _gather
     Tensor.all = _all
     Tensor.any = _any
